Revert "move utils into own folder - add sep files"
This reverts commit 0cd541cb6c9f95f7c42918ad0dd85d41f75c9461.
