+++ eq-author-api/schema/resolvers/logic/binaryExpression2/index.js
@@ -19,7 +19,7 @@ const {
   getSkipConditions,
   getAllExpressionGroups,
   returnValidationErrors,
-} = require("../../utils/utils");
+} = require("../../utils");
 
 const Resolvers = {};
