+++ eq-author-api/schema/resolvers/logic/routing2/expressionGroup2/index.js
@@ -3,7 +3,7 @@ const Resolvers = {};
 const { find, flatMap, get } = require("lodash/fp");
 const { createMutation } = require("../../../createMutation");
 
-const { getPages } = require("../../../utils/utils");
+const { getPages } = require("../../../utils");
 
 Resolvers.ExpressionGroup2 = {
   expressions: (expressionGroup) => expressionGroup.expressions,
