+++ eq-author-api/schema/resolvers/logic/routing2/routing2/index.js
@@ -12,11 +12,7 @@ const {
   createLeftSide,
 } = require("../../../../../src/businessLogic");
 
-const {
-  getPages,
-  getPageById,
-  getRoutingById,
-} = require("../../../utils/utils");
+const { getPages, getPageById, getRoutingById } = require("../../../utils");
 
 const isMutuallyExclusiveDestination = isMutuallyExclusive([
   "sectionId",
