+++ eq-author-api/schema/resolvers/logic/routing2/routingRule2/index.js
@@ -18,7 +18,7 @@ const {
   getRoutingById,
   getRoutingRuleById,
   returnValidationErrors,
-} = require("../../../utils/utils");
+} = require("../../../utils");
 
 const isMutuallyExclusiveDestination = isMutuallyExclusive([
   "sectionId",
