+++ eq-author-api/schema/resolvers/pages/questionPage.js
@@ -7,7 +7,7 @@ const {
   getSectionByPageId,
   getFolderById,
   getFolderByPageId,
-} = require("../utils/utils");
+} = require("../utils");
 
 const { createQuestionPage } = require("../../../src/businessLogic");
