+++ eq-author-api/src/validation/customKeywords/validateLeftHandSide.js
@@ -8,9 +8,7 @@ const {
 const createValidationError = require("../createValidationError");
 
 const { getPath } = require("../utils");
-const {
-  getAbsolutePositionById,
-} = require("../../../schema/resolvers/utils/utils");
+const { getAbsolutePositionById } = require("../../../schema/resolvers/utils");
 
 module.exports = function (ajv) {
   ajv.addKeyword("validateLeftHandSide", {
